+++ src/entities/character/Oneal.java
@@ -5,7 +5,7 @@ import graphics.Sprite;
 
 public class Oneal extends Enemy {
 
-    public Oneal(int x, int y, double velocityX, double velocityY, Sprite sprite) {
+    public Oneal(int x, int y, int velocityX, int velocityY, Sprite sprite) {
         super(x, y, velocityX, velocityY, sprite);
         animatedSprites.put("LEFT", new Sprite[]{Sprite.oneal_left1, Sprite.oneal_left2, Sprite.oneal_left3});
         animatedSprites.put("RIGHT", new Sprite[]{Sprite.oneal_right1, Sprite.oneal_right2, Sprite.oneal_right3});
